localStorage에 데이터 저장 및 업데이트 기능 구현
