+++ src/js/utils/storage.js
@@ -1,0 +1,9 @@
+const storage = window.localStorage;
+
+export const getStorage = () => {
+  return JSON.parse(storage.getItem('items')) || [];
+}
+
+export const setStorage = (key, value) => {
+  storage.setItem(key, JSON.stringify(value));
+}
